requirements.txt
functions-framework==3.*
requests==2.*
google-auth==2.*
google-auth-httplib2==0.*
google-api-python-client==2.*
google-cloud-secret-manager==2.*
Pillow==10.*
python-dotenv==1.*
